Migrating components to TS.
